refactor: useFetchStory hook을 useStory로 이동
