+++ src/pages/Story.tsx
@@ -4,7 +4,7 @@ import { Box, CircularProgress, Divider } from '@mui/material';
 import StoryComment from '../components/Story/StoryComment';
 import StoryInfo from '../components/Story/StoryInfo';
 import { TOKEN_KEY } from '../constants/auth';
-import useFetchStory from '../hooks/useFetchStory';
+import { useFetchStory } from '../hooks/useStory';
 import { getLocalStorage } from '../utils/storage';
 
 const Story = () => {
